config files name update

+++ Team-service/src/main/java/com/fpl/repository/TeamRepository.java
@@ -2,7 +2,7 @@ package com.fpl.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.fpl.service.Team;
+import com.fpl.entity.Team;
 
-public interface TeamRepository extends JpaRepository<Team, Long> {
+public interface TeamRepository extends JpaRepository<Team, Integer> {
 }
